Ytelser: konverterer dager til uker
